Converted the Database to a singleton,implement in the UserService the registerUser function,and implement in the AuthController the handleRegister function,so we can insert a user in our database,but we don't validate the user yet.

+++ database.sql
@@ -11,7 +11,7 @@ DROP TYPE IF EXISTS "status" CASCADE;
 CREATE TABLE users(
     user_id SERIAL Primary Key,
 	user_email VARCHAR(30) UNIQUE,
-    user_taxnum int UNIQUE,
+    user_taxnum BIGINT UNIQUE,
 	user_password TEXT,
 	confirmed bool 
 );
